Auto-equip highest level hatchet upon breaking
Automatically equips the highest level hatchet upon breaking - if none are available, the woodcutters may work with 'no hatchet', which has an accuracy of 0.

For cutting a tree at level 99, there is a 100% chance to cut a regular tree, a 33% chance to cut a willow tree, a 9% chance to cut a magic tree, and a 6% chance to cut an elder tree.
At level 1, there is a 5% chance to cut a regular tree with no hatchet. I think this is a reasonable compromise.